// analysis of software requirements/aswr5/src/app/models/enums.ts
export enum DataType {
  Event = 'Event',
  Risk = 'Risk',
}

export enum RiskType {
  tech = 'Tech',
  cost = 'Cost',
  plan = 'Plan',
  realisation = 'Realisation',
  total = 'Total',
}

export enum Type {
  Evaluation = 'Evaluation',
  Probability = 'Probability',
}

export enum Status {
  VeryLow = 'Дуже низький',
  Low = 'Низький',
  Medium = 'Середній',
  High = 'Високий',
  VeryHigh = 'Дуже високий',
}